+++ tools/fs/files.js
@@ -12,7 +12,7 @@ var util = require('util');
 var _ = require('underscore');
 var Fiber = require('fibers');
 var crypto = require('crypto');
-var spawn = require("child_process").spawn;
+const { spawn, execFile } = require("child_process");
 
 var rimraf = require('rimraf');
 var sourcemap = require('source-map');
@@ -127,6 +127,29 @@ files.findPackageDir = function (filepath) {
   };
 
   return findUpwards(isPackageDir, filepath);
+};
+
+// Returns the hash of the current Git HEAD revision of the application,
+// if possible. Always resolves rather than rejecting (unless something
+// truly unexpected happens). The result value is a string when a Git
+// revision was successfully resolved, or undefined otherwise.
+files.findGitCommitHash = function (path) {
+  return new Promise(resolve => {
+    const appDir = files.findAppDir(path);
+    if (appDir) {
+      const proc = execFile("git", ["rev-parse", "HEAD"], {
+        cwd: files.convertToOSPath(appDir),
+      }, (error, stdout) => {
+        if (! error && typeof stdout === "string") {
+          resolve(stdout.trim());
+        } else {
+          resolve();
+        }
+      });
+    } else {
+      resolve();
+    }
+  }).await();
 };
 
 // create a .gitignore file in dirPath if one doesn't exist. add