+++ scripts/build-dev-bundle-common.sh
@@ -8,7 +8,7 @@ ARCH=$(uname -m)
 NODE_VERSION=8.15.0
 MONGO_VERSION_64BIT=4.0.4
 MONGO_VERSION_32BIT=3.2.21
-NPM_VERSION=6.5.0
+NPM_VERSION=6.8.0
 
 # If we built Node from source on Jenkins, this is the build number.
 NODE_BUILD_NUMBER=