+++ History.md
@@ -69,6 +69,13 @@ N/A
   [Issue #10393](https://github.com/meteor/meteor/issues/10393)
   [PR #10414](https://github.com/meteor/meteor/pull/10414)
 
+* If your application is using Git for version control, the current Git
+  revision string will be exposed via `Meteor.gitRevision` while the app
+  is running, and also via the `"gitRevision"` property in the `star.json`
+  file located in the root directory of builds produced by `meteor build`.
+  If you are not using Git, neither property will be defined.
+  [PR #10442](https://github.com/meteor/meteor/pull/10442)
+
 ## v1.8.0.2, 2019-01-07
 
 ### Breaking changes